+++ gnu/fsmeter.cc
@@ -2,6 +2,8 @@
 //  Copyright (c) 2015
 //  by Mike Romberg ( mike-romberg@comcast.net )
 //
+//  This file may be distributed under terms of the GPL
+//
 #include "fsmeter.h"
 #include "fsutil.h"
 #include "strutil.h"
@@ -50,7 +52,7 @@ void FSMeter::checkevent( void ) {
         setBGColor(_bgColor);
 
         // free, total (for root)
-        std::pair<size_t, size_t> fsSize = util::fs::getSpace(_path, true);
+        std::pair<uint64_t, uint64_t> fsSize = util::fs::getSpace(_path, true);
 
         logDebug << _path << ":\t" << fsSize << std::endl;
